+++ fixtures/app_with_framework_where_framework_failing_swiftlint/App/ClassA.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+public class ClassA {
+    public init() {}
+}
+