+++ fixtures/ios_app_with_framework_and_resources/App/Sources/AppDelegate.swift
@@ -9,7 +9,7 @@ import StaticFramework5
 @UIApplicationMain
 class AppDelegate: UIResponder, UIApplicationDelegate {
     var window: UIWindow?
-    let staticFrameworkResources = StaticFrameworkResouces()
+    let staticFrameworkResources = StaticFrameworkResources()
     let staticFramework2Resources = StaticFramework2Resources()
     let resourcesStaticFramework3 = ResourcesStaticFramework3()
     let resourcesStaticFramework4 = ResourcesStaticFramework4()
@@ -18,11 +18,16 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         let framework1 = Framework1File()
 
         print(hello())
+        print("Custom plist: \(Environment.myKey)")
         print("AppDelegate -> \(framework1.hello())")
         print("Main bundle image: \(String(describing: UIImage(named: "tuist")))")
+        print("Asset catalogue image: \(String(describing: AppAsset.assetCatalogLogo.image))")
+        print("Strings: \(AppStrings.Greetings.morning)")
+        print("Font: \(AppFontFamily.SFProDisplay.bold.name)")
         print("StaticFrameworkResource image: \(String(describing: staticFrameworkResources.tuist))")
         print("StaticFramework2Resource image: \(String(describing: staticFramework2Resources.loadImage()))")
         print("StaticFramework3Resource image: \(String(describing: resourcesStaticFramework3.loadImage()))")
+        print("StaticFramework3Resource asset catalogue image: \(String(describing: StaticFramework3Asset.assetCatalogLogo.image))")
         print("StaticFramework4Resource image: \(String(describing: resourcesStaticFramework4.loadImage()))")
         print("StaticFramework5Resource image: \(String(describing: UIImage(named:"StaticFramework5Resources-tuist", in: StaticFramework5Resources.bundle, compatibleWith: nil)))")
     }